requirements.txt
blockcypher==1.0.93
PyQt6==6.3.0
python-dotenv==0.20.0
